refactor(cosmo_val): extract PseudoClMixin (pseudo_cl)
